Report ambiguous owners of JVM types (#13319)
Ambiguous providers of types are common in some of our sample repositories, and so #13024 is important to avoid overly broad graphs, and to provide guidance in ambiguous situations.

The added test shows that when two providers of a type are present, neither are added unless they are disambiguated.

This change also removes package-awareness from `PackageRootedDependencyMap`, because we believe that only thirdparty packages or explicitly detected wildcard imports should use fuzzy lookups for packages. @tdyas will be adding a thirdparty-specific data structure.

Fixes #13024.