Package local dists into .pex files. (#13306)
Previously if you depended on a python_distribution target we would build it locally
and install it into the chroot, so you could consume dists locally during development
without publishing them. This is particularly useful for native extensions, that cannot
be consumed directly from source.

However this support only worked for run, repl and test. If you used package to
build a PEX, that PEX wouldn't have that dist embedded in it. You would have had to
publish the dist and then consume it via a requirement.

This change adds the missing support for embedding a locally-built dist into a PEX.
Now the local dist support is uniform. Note that you can still publish and consume via
requirement - in that case don't have your pex_binary depend on the python_distribution.
But if it does, the implication is that you want to build it locally (before this PR there was
no sensible meaning to a pex_binary depending on a python_distribution, so no semantics
have changed).

[ci skip-rust]

[ci skip-build-wheels]