Move `[python] interpreter_search_paths` to `[python-bootstrap] search_path`, and add support for sandboxed Python scripts in `@rule`s (#13340)
Bash scripting has proved useful for quick sandboxed tasks accomplished by `@rules`. This change exposes Python >=3.6 in a similar fashion from the core of Pants (i.e., without the Python backend installed), for basic scripts without thirdparty dependencies.

Python also has the advantage of a large standard library, which is used immediately in #13335, to expose `gunzip` in environments without those CLI binaries. Eventually some usecases for sandboxed binaries like this may be ported to purpose-built Rust binaries, but it is safe to assume that there will always still be scripts.

This change exposes a `PythonBinary` (akin to `BashBinary`). To do so, it moves the interpreter-search-path and interpreter bootstrap name options into a core `Subsystem`: `python-bootstrap`.

[ci skip-rust]
[ci skip-build-wheels]