A PipRequirement class to replace most of our use of pkg_resources.Requirement (#13348)
We use pkg_resources.Requirement all over the codebase to represent a user requirement string (e.g., from requirements.txt or an explicit python_requirement target).

These requirements are expected by users to be "any (non-editable perhaps) requirement that Pip understands". But that set is a superset of what pkg_resources.Requirement understands. E.g., Pip has a syntax for old-style (pre-PEP 440) VCS requirements, and Pip requirements can also take inline options, such as --hash.

To prepare for being able to handle such requirements in the future, this change introduces a PipRequirement class, which is a drop-in replacement for pkg_resources.Requirement.

It also drops this class in the various places where it makes sense to do so.

Note that PipRequirement currently has no capabilities beyond delegating to pkg_resources.Requirement. But it will serve as a single locus for adding better requirement support in future changes.

[ci skip-rust]

[ci skip-build-wheels]