assume version 1 for older lockfiles without `version` key (#13399)
As reported in Slack, Python lockfiles created in Pants v2.7.x do not have a `version` key in the lockfile metadata. Pants v2.8.x generates lockfiles with the `version` key in the metdata and expects the files to contain that key. It had no fallback for older lockfiles without the `version` key. This broke an upgrade for one user from v2.7.0rc2 to v2.8.0rc0. 

Solution: Pants should just assume version 1 if it does not find the `version` key in lockfile metadata.

[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]