Use `--layout=packed` for all monolithic resolves. (cherrypick of #13400) (#13403)
As described in #13398, we should use `--layout=packed` for all monolithic resolves (lockfiles, all_constraints, etc), in addition to for `internal_only` resolves.

Note that this is _not_ sufficient to allow for cache reuse across internal PEXes (i.e. those created for tests) and external PEXes (`pex_binary`), because the `repository.pex` is constructed with other differing arguments. 

Fixes #13398.

[ci skip-rust]
[ci skip-build-wheels]