Don't error for invalid Go third-party packages unless building them (Cherry-pick of #13405) (#13410)
Closes https://github.com/pantsbuild/pants/issues/13391. 

We now consistently use the philosophy implemented in https://github.com/pantsbuild/pants/pull/13377 with third-party packages for Go. So long as that package has a valid `import_path` (its unique identifier), we only error when building the package. We don't error with target generation and using it for project introspection. 

This is particularly important for Go because we generate a target for every package from third-party modules, even if it will never be used. For example, the failure from https://github.com/pantsbuild/pants/pull/13377 was for `github.com/cncf/udpa/test/build`, i.e. a test package never used by first-party code.

With this fix, both Helm and external-dns now work with `./pants list ::`. (Not sure about `check` because performance is too slow to get a result...)

[ci skip-rust]
[ci skip-build-wheels]