Fixes to interpreter constraint computation.
When computing constraints for a target set, previously we'd return
empty constraints if the target set had no intepreter_constraints fields.
Instead, we should have been using the global constraints, so that
the behavior is continuous (each interpreter_constraints field adds
constraints, so no fields should be the least constrained). So we now
do so.

Also, implements a snap_to_minimum() function that returns constraints
representing the minimum version supported by the original constraints.
Reimplements minimum_python_version() to use this snapped version.

Also, sets a more generous upper bound for possible patch versions of
Python, since there's no harm in doing so, and risk in not doing so.

[ci skip-rust]

[ci skip-build-wheels]