[Internal] Java & Scala: Supply only dependencies to the compiler when building (#13447)
Adds `FilterDependenciesRequest`, which takes a Coursier lockfile and a set of coordinates, and returns a cut-down version of that lockfile that contains only the specified coordinates, and if `transitive` is set, the transitive dependencies of those coordinates. This means that targets can share a resolve, but only the artifacts needed to build a given target are supplied to the compiler. This should improve build performance.

nb. `transitive` is set to `True` by default until configuration is added.

This resolves #13366.