Go: improve performance of first-party package analysis (#13476)
## Motivation

`go list` requires that a package's dependency graph be available to analyze a package, even if all we want to do is obtain the package's imports which can be found by just parsing sources. This is fine if the packages are already available on the GOPATH. With the Pants execution sandbox, however, this means that Pants has to copy large trees of sources into the input root to make those packages available even if not actually needed. And that copying into the input root has a significant performance impact.

## Solution

Vendor some of the package analysis code from the [go/build standard library package](https://cs.opensource.google/go/go/+/refs/tags/go1.17.2:src/go/build/build.go) into Pants. This will be the basis for not using `go list` for package analysis.

Interestingly, Bazel rules_go went in a similar direction for package analysis.

## Benchmark

Running in https://github.com/toolchainlabs/remote-api-tools.

Before:

```
❯ hyperfine -r 5 './pants_from_sources --no-process-execution-local-cache --no-pantsd dependencies ::'
  Time (mean ± σ):     19.742 s ±  0.801 s    [User: 13.908 s, System: 26.641 s]
  Range (min … max):   19.113 s … 21.081 s    5 runs
```

After:

```
❯ hyperfine -r 5 './pants_from_sources --no-process-execution-local-cache --no-pantsd dependencies ::'
  Time (mean ± σ):     14.472 s ±  1.107 s    [User: 9.550 s, System: 5.793 s]
  Range (min … max):   13.561 s … 16.101 s    5 runs
```

Note the dramatic drop in System time. We suspect that this was resulting in IO contention before, which was causing Pants to hang in larger repositories like Helm.

[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]