[internal] Make explicit when using the GIL with `store_*` FFI helpers (#13516)
See https://github.com/pantsbuild/pants/pull/13515 for the motivation. Making GIL access explicit does make call sites more verbose, but it also makes it easier to reason about where the GIL is used and ensures we are not continually dropping and then reacquiring the GIL. 

Compare to https://github.com/pantsbuild/pants/pull/11186, which resulted in sharing the GIL but also relied on an intricate implementation detail. The modeling is now explicit.