[internal] Inline some of `externs/mod.rs` and make GIL acquisition explicit (#13522)
A couple of refactors to simplify our rust-cpython FFI code. 

Most notable is rewriting `gettattr_as_string` to be `getattr_as_optional_string`, which is always how it was behaving. That is, we used to tolerate if Python was set to `str | None`, even though most call sites expected `str`. This rename makes clear the difference.