Add support for multi-language JVM compilation (#13519)
In order to support multi-language JVM compilation (`Java` and `Scala` as dependencies of one another), add a `ClasspathEntryRequest` `@union`, which is used to install a set of `@rules` which can provide a `ClasspathEntry` for a resolve and a `CoarsenedTarget` containing known `FieldSet` types.

The `ClasspathEntryRequest` `@union` also allows for non-compilation usecases, and is used here by `Coursier` to expose `ClasspathEntry`s for `jvm_artifact` targets. In future, it will also be used to expose `resources` in #13521.

Additionally, `ClasspathEntry` becomes a recursive structure, which allows `ClasspathEntryRequest` instances to define the entire classpath layout, and decide what to export vs leave as dependencies. Also reintroduces the `filename` to `ClasspathEntry` in order to restore classpath ordering (oops).

Fixes #13329. See #13520 for followup work to allow for compiling cycles/`CoarsenedTarget`s containing both `Java` and `Scala`.