[internal] go: support passing embedcfg to compiler for file embedding (#13509)
Add `EmbedConfig` type representing the configuration passed to the Go compiler to configure file embedding. This is preliminary work to support the Pants rules that will resolve which files to embed.

The configuration uses a two-level mapping from pattern to actual filesystem paths with content. The first mapping is via the `patterns` dictionary which maps the file glob patterns used in `//go:embed` directives to a list of relative paths (relative to the package directory) which are used to define a directory hierarchy when the embedding variable has the `embed.FS` type. Those relative paths are mapped to actual filesystem paths for content via the `files` dictionary.

[ci skip-rust]