[jvm] Add support for consuming types from the unnamed package. (#13536)
The assertion in #13214 was technically true: you cannot `import` a symbol from the unnamed package. But a file may still use a symbol from the unnamed package if it is also in the unnamed package.

Although heavily using the unnamed package is ill-advised, we should account for the fact that packages are optional. 

[ci skip-rust]
[ci skip-build-wheels]