[internal] scala: extract imports from source files (#13556)
Extract imports from Scala sources and report them for dependency inference. Unlike Java, Scala imports can occur in many different scopes in a file (and not just the file's top-level scope), and so this PR reports the imports by scope. (I am not sure if reporting by scope name will be useful but easy enough to extract that information now so I can attempt to make use of it in subsequent PRs.)

[ci skip-rust]