Do not invalidate parent paths when only file content has changed (#13566)
We have avoided fine grained interaction with the `notify` crate so far, but now that we've seen that `notify` is reliable, we can slowly introduce more logic around event types.

This change skips invalidating the parent of a path when the event type is file-data-only (`Modify(Data(Content))`). As mentioned in #11707 though, it's not as effective as we might have hoped, since editors tend to write to temp files and then rename them. It might still have a positive impact in some cases, and seems like a relatively safe first step.

Fixes #11707.