[jvm] UX improvements for compilation and thirdparty dependency lookups (#13573)
Nailgun processes were not wrapped in a workunit, which resulted in their entire runtime being blamed to the `Scheduling: $description` process wrapper workunit. Add a workunit, and additionally improve the process description for Java analysis.

Additionally, improve the error when the coordinate from a `jvm_artifact` is not present in a lockfile, by suggesting updating the lockfile.

Fixes #13364.