[plugin-api] Add `default_glob_match_error_behavior` to `SourcesField`. (#13578)
This optional field allows specifying that a default glob value for a source field is optional (i.e. should not warn or err in case the glob does not match any files).

fixes #13567

There was a bug in the `SingleSourceField.path_globs()` that was also fixed in this PR.
If the field has a default value then that would use the wrong glob expansion conjunction when your provided source value would have the same set of characters as the default value.
For example, a default value of `"file"` and a provided value of `"life"` would treat this as if it was the default value was provided.
Note however, that with a single glob, the conjunction used does not matter (but when using the new default glob error override introduced in this PR, this distinction does matter).