Add `instructions` field to `docker_image` to support generating the Dockerfile (#13386)
In order to support building Docker images without a dedicated `Dockerfile` on disk,  the `instructions` field may be used to  create the `Dockerfile` on demand.

This is useful for instance when you have lists of images that you want to mirror (i.e. for air gapped environment) in text files, using a plugin that can turn these files into targets similar to how `python_requirements` consumes `requirements.txt` files. This helps keep your project clean and DRY.

Example:
```python
docker_image(
  name="demo_image",
  instructions=[
    "FROM baseimg:tag",
    "COPY src /dest/",
    "ENTRYPOINT demo",
    ...
  ]
)
```

fixes #13360 


## Alternative implementations

A dedicated target `dockerfile` was considered, but proved to be unnecessarily complicated to make intuitive to work with. See the original proposal in #13360.