JVM/Java: Add Export types to dependency inference code and include exports in compile results for Java sources (#13603)
This adds the concept of "export" types to our JVM dependency analysis. An export type is any type that is required to fulfil the public or inheritable interface of a given type.

This includes:
* Superclasses by `extends` or `implements`
* Return values of `public` or `protected` methods
* Parameter types for polymorphic methods

At the moment, the actual implementation is coarser than necessary: we export _all_ return values and _all_ parameter types, but these could be implemented to be closer to the above definition.

Export types are required by the Java compiler to verify whether any a subclass is valid, or a given method call is possible. Importantly, an export type depends on the transitive closure of all export types.

To achieve this, exports are implemented on the Pants side by merging the `ClasspathEntry` objects of a compilation unit with the `ClasspathEntry` of each export type. This ensures that the relevant hierarchy of classes are present for the compiler.