[internal] Refactor finding owning `go_mod` for first-party packages (#13695)
Prework for https://github.com/pantsbuild/pants/issues/13488.

We used to find the owning `go_mod` by relying on the fact that `go_package` targets are generated and simply converting the address to its target generator. That won't work once we allow manually creating `go_package` targets, though.

We could require that a `go_package` target have a `go_mod` field that points to the address of the owning `go_mod`. That would avoid the need for this rule and be slightly faster. But that is not ergonomic. Even with `tailor` setting it, it would not be resilient to reorganizing files and it would add extra boilerplate.

[ci skip-rust]
[ci skip-build-wheels]