[internal] tailor adds go_package targets (#13703)
Follow up to https://github.com/pantsbuild/pants/pull/13702.

## Weird edge when upgrading from Pants 2.8: `go_binary`

`go_binary` has a `main` field that points to a `go_package`, but usually it's left off to default to the `go_package` defined in the directory.

Because we no longer generate `go_package` targets, the `go_binary` will now not have a `go_package` target, so our `tailor` rule will error (because it tries to determine which main packages are already owned).

I tried to fix this by making our `GoBinaryMainPackage` fallible, but it was pretty complex code and it resulted in trying to add a second `go_binary` target because it looked like the package was unowned. We could make that much more nuanced, like check if the `main` field is set for the `go_binary` w/ a missing `go_package`. But the code was getting really complex for what should be a one-time migration issue with an experimental backend.

So I kept it simple and made the error message more informative.

[ci skip-rust]