Fix UI rendering of Go packages with no tests (#13710)
Before, we would confusingly render a package without tests as having succeeded, even though it did not run.

This reverts https://github.com/pantsbuild/pants/pull/12516. It turns out that while we can eagerly opt out of running tests sometimes based only on the Target API, e.g. looking at the `skip_tests` field, we sometimes need more information. With Go, we must analyze the package to see if it has tests.

If the analysis didn't succeed to determine if we have tests, we pessimistically assume there were tests and render the analysis/compilation failure as a test failure.

[ci skip-rust]
[ci skip-build-wheels]