Introduce OptionalPex/OptionalPexRequest.
So we can operate monadically in situations where there
may or may not be an underlying Pex.

Utilizes this in one such case we currently have - requesting
a repository pex.

This brings repository pexes more inline with other pexes - you
request them via a PexRequest, except you must also account for
the possibility that no repository pex can be used.

[ci skip-rust]

[ci skip-build-wheels]