Add support for inferring relative imports from Scala (#13775)
Relative imports are common in Scala, and not supporting them caused ~12% of the issues in our test repository. See #13740 for background.

To support them, compares the final token of an import (or its alias, if it has one) to the first token of any multi-token symbol that is scoped below the import.

Fixes #13740.