An option to test/run/repl against the entire lockfile. (#13732)
This cuts out the subsetting, which can add a lot of time
to runs, and can also bloat the cache dramatically.

One downside is less hermetic runs - you may be incidentally
relying on requirements that you don't depend on but happen to
be present in the lockfile (but this seems like less
of an issue with dep inference).

A more significant downside is that any changes to the lockfile
will invalidate all tests. Still, this is a performance knob that
users can decide how to use. In many cases that may be a very
acceptable consequence.

[ci skip-rust]

[ci skip-build-wheels]