[internal] DRY loading internal Go binaries (#13800)
Follow up to https://github.com/pantsbuild/pants/pull/13791. 

We soon need to add a third Go binary. This DRYs that setup.

It also uses a distinct folder for each Go binary, which makes the code easier to understand and will allow us to more easily add tests in the future.

[ci skip-rust]
[ci skip-build-wheels]