[jvm] Spawn nailgun servers outside the pool lock (#13796)
As described in #13788, when we spawn nailgun processes, we currently do so inside the pool's lock. This blocks other processes from being acquired or started until the new server has started.

To resolve this, `PoolEntry`s now hold an `Option<NailgunProcess>` instead, which accounts for the fact that although they are eagerly started, the entry is potentially visible to other pool consumers before it has fully started.

Fixes #13788.