[internal] Clean up Go `embed` support's handling of dependencies (#13801)
Before, we were both analyzing embed patterns and using that to create the embed config in the same program. The issue is that creating the embed config requires having resource files present in the chroot, so finding the dependencies of the `go_package`. But because the rule to determine `FirstPartyPkgInfo` is used by dependency inference for the `go_package`, looking at the dependencies of that target to get `resources` results in a graph cycle.

The robust fix is to split out creation of the embed config into a distinct program. Concretely, `FirstPartyPkgInfo` is now `FirstPartyPkgAnalysis` and `FirstPartyPkgDigest`.

[ci skip-rust]