[internal] jvm: fix default paths for tool lockfiles (#13822)
Fix how the `--TOOL-lockfile` option's default is set. It was set to a default of its location in the Pants repository and not the special string `<default>`. The special string should trigger using the resource for the default lockfile embedded in the Pants distribution. Pants will error if the user explicitly requests resolving the lockfile of a tool using the default lockfile.

In the Pants repository itself, the `--TOOL-lockfile` options are set appropriately for their locations in the Pants repository (which are then consumed by `resource` targets).

This PR only fixes the current broken behavior, preserving the status quo.

[ci skip-rust]