scala: add `scalafmt` support for `fmt` and `lint` goals (#13814)
Add the `scalafmt` tool for `fmt` and `lint` goals.

Pants will choose the nearest ancestor `.scalafmt.conf` file as the configuration file to use for a source file. This differs from the logic used by `scalafmt` which only chooses between a `.scalafmt.conf` in the same directory as a file or one in the root of the repository. This should provide a better experience for multi-team monorepos where it could be useful to have a `.scalafmt.conf` that applies to a subtree of the repository.

Closes #11943.