JVM: Adds `jar` field to `jvm_artifact` (#13834)
Adds the `jar` field to `jvm_artifact`, which allows users to check a JAR file into the repository, and use that to fulfil dependencies as needed. This is because sandboxing is generally incompatible with file URLs, so we make use of Pants' existing filesystem support.

This requires a small bit of weirdness -- firstly, Coursier requires absolute file URLs, which Pants' sandboxing doesn't really like, so so `coursier_wrapper.sh` now preprocesses any args containing a placeholder to use the actual current working directory. Secondly, `ArtifactRequirements` are now preprocessed to check if a requirement uses a local JVM artifact, and builds a digest creating those JARs and sets the URL to be the file URL with the working directory placeholder.

This closes #13820