[internal] Intern using a Python dict, and remove reverse mapping (#13859)
Profiling showed that we're still spending quite a bit of time acquiring and releasing the GIL for interning (since we release it to interact with the `HashMap`, and then reacquire it for `Eq`). But it also showed a surprising amount of time spent interacting with the locks around the internal `HashMap`s.

Since we're already under the GIL, and since Python dicts are plenty optimized on their own, this change moves interning to directly using a `PyDict` (effectively `dict[Any,int]`). It also removes the reverse mapping by giving `Key` a reference to its `Value`. Finally, it replaces the `TODO` about removing the `Arc` in `Value` with an explanation that using `Arc<PyObject>` is about 4% faster.

`./pants dependencies --transitive ::` gets ~6% faster.

[ci skip-build-wheels]