Fix `.pyi` files not automatically depending on `__init__.py` (Cherry-pick of #13844) (#13858)
Closes https://github.com/pantsbuild/pants/issues/13654. We were skipping automatically including `__init__.py` because we incorrectly short-circuited if a source file ended in `.pyi` instead of `.py`.

We also were not automatically adding dependencies on `__init__.pyi` files.

Finally, this simplifies `AncestorFilesRequest` to take `tuple[str, ...]` as input files, rather than `Snapshot`. That allows us to avoid having to call `HydrateSources` for the `conftest` and `__init__` dependency inference rules. We can simply look at the target's source field to determine what the input file is, which is faster.

[ci skip-rust]
[ci skip-build-wheels]

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>