[internal] scala: refactor `scalapb` codegen and add JVM `protoc` plugin support (#13873)
Refactor the ScalaPB protobuf codegen support to use a new compiled shim for invoking ScalaPB instead of passing the ScalaPB `protoc` plugin to the `protoc` binary. This will allow us to be as close in behavior to the `scalapbc` standalone compiler as is feasible, while still allowing for Pants-specific customizations. (It also allows nailgun to cache the shim, unlike a solution which would invoke the GraalVM-built native binary for `scalapbc`.)

Add support for JVM-based `protoc` plugins. The plugins are exposed via the `--scalapb-jvm-plugins` option. The integration test demonstrates enabling code generation with the https://github.com/typelevel/fs2-grpc gRPC code generator.