[internal] Fix `resolve` field handling for JVM (#13876)
We weren't properly handling the `resolve` field, which blocked switching our test targets to use `resolve` instead of `compatible_resolves` in https://github.com/pantsbuild/pants/pull/13870.

* When generating lockfiles, we ignored targets with `resolve`.
* When determining which lockfile to use, we ignored `resolve`.

The algorithm to determine which lockfile to use is fairly simplistic. If for the set of targets (~`CoarsenedTargets`) there is a `resolve` field, use that. Else, use any of the `compatible_resolves`. 

Some followups we still need to do:

* validate that the input to `CoursierResolveKey` is all compatible with each other
* validate that the dependencies of the `CoursierResolveKey` targets are compatible
* always set a default resolve
* add `default_compatible_resolves`
* update check to run on all relevant resolves for the root, whereas now we choose one