[internal] JVM only infers dependencies on artifacts in the target's resolve(s) (#13907)
Closes https://github.com/pantsbuild/pants/issues/13621.

First-party targets like `java_source` will only infer dependencies on `jvm_artifact`s from one of their `compatible_resolves` / `resolve` field. This works around issues with ambiguous dependency inference.

Note that we do not eagerly error if you explicitly depend on a `jvm_artifact` not compatible with the target's resolve(s). We may want to do that in a followup.

[ci skip-rust]