[jvm] Apply the resolve for `deploy_jar` and restore validation of resolve compatibility (#13911)
The `resolve` field for `deploy_jar` was not being respected, because the `deploy_jar` itself was not included in the targets for which we were calculating the resolve (only its dependencies). To fix this, we add support to `ClasspathEntryRequest` for "root only" requests, which allow for requesting a classpath for a target when it is a root, but failing when the target type is used as an inner node (since we don't want `deploy_jar`s to be used as dependencies of JVM compile targets anytime soon).

Additionally, restore the validation of resolve compatibility when computing a `CoursierResolveKey` (temporarily removed in #13876). This time around, the `@rule` takes `CoarsenedTargets` as input, which removes redundant walking while checking the transitive graph, while still allowing us to differentiate between the roots of the request, and the dependencies of the request.

[ci skip-rust]
[ci skip-build-wheels]