Fix `tailor` not setting the `name` field (Cherry-pick of #13913) (#13924)
We were setting the target name to always be `docker`, but not setting the kwargs so the name _actually_ ended up being the default of the directory name. This caused tailor to not properly handle collisions in the target name.

This fixes it by removing the gotcha entirely. Now, `core/goals/tailor.py` determines if the `name` field will be set or not based on whether the `name` is the default of the directory name.

[ci skip-rust]
[ci skip-build-wheels]