Revert "Turn off validation for unset `SourcesField" if `required=False` (#13856) (#13944)
We need to override how `SIngleSourceField` sets `expected_num_files = 1` because that causes issues when the `jar` field is unset: https://github.com/pantsbuild/pants/issues/13851. One approach is to use `expected_num_files = range(0, 2)`, but instead https://github.com/pantsbuild/pants/pull/13856 turned off the validation entirely.

As explained in that ticket, we do still want to validate that the file is exactly one if the field was explicitly set.

There's another way to do that though. We already validate that source globs are valid via `--files-not-found-behavior`. If you explicitly set the value, then we will set up `PathGlobs` such that it will warn or error if there is no match, meaning we handle zero files. Because we ban `*` and also have `expected_num_files`, we already handle >1 file.

[ci skip-rust]
[ci skip-build-wheels]