[internal] Rename `[python].experimental_resolves_to_lockfiles` to `[python].experimental_resolves` (#13950)
We've been using `[jvm].resolves` and I think it works well. Turns out we reference the option a lot in `help` messages etc. The `to_lockfiles` part doesn't pull its weight.

Because we marked this option as highly experimental and it didn't even really work properly, we don't use a deprecation. 

[ci skip-rust]
[ci skip-build-wheels]